Improve class augmentation for Element
