+++ packages/alfa-dom/src/node/element.ts
@@ -269,113 +269,19 @@ export class Element<N extends string = string>
   }
 
   /*
-   * This collects methods that are specific to some kind of elements.
-   * Note that they can be used in the predicates that are included in the
-   * namespace later in this very file, as long as the namespace is declared
-   * after the class (so the references don't go full circle).
+   * This collects caches for methods that are specific to some kind of elements.
+   * The actual methods are declared in element/augment.ts to de-clutter this
+   * class. However, we need to declare the variables here as they would
+   * otherwise be public which exposes too much. And they must be `protected`
+   * to be accessible by the augments.
    */
 
-  private _inputType: helpers.InputType | undefined;
-
-  /**
-   * {@link https://html.spec.whatwg.org/#attr-input-type}
-   */
-  public inputType(this: Element<"input">): helpers.InputType {
-    if (this._inputType === undefined) {
-      this._inputType = this.attribute("type")
-        .flatMap((attribute) =>
-          attribute.enumerate(
-            "hidden",
-            "search",
-            "tel",
-            "url",
-            "email",
-            "password",
-            "date",
-            "month",
-            "week",
-            "time",
-            "datetime-local",
-            "number",
-            "range",
-            "color",
-            "checkbox",
-            "radio",
-            "file",
-            "submit",
-            "image",
-            "reset",
-            "button",
-            "text",
-          ),
-        )
-        .getOr("text");
-    }
-    return this._inputType;
-  }
-
-  private _displaySize: number | undefined;
-
-  /**
-   * {@link https://html.spec.whatwg.org/multipage/form-elements.html#concept-select-size}
-   *
-   * @remarks
-   * The size IDL attribute should have a value of 0, not 1 or 4, when the
-   *   content attribute is undefined. This is for historical reasons. In our
-   *   case, this is not affecting the results and it is easier to treat it as
-   *   the actual displayed size.
-   * {@link https://html.spec.whatwg.org/multipage/form-elements.html#dom-select-size}
-   */
-  public displaySize(this: Element<"select">): number {
-    if (this._displaySize === undefined) {
-      this._displaySize = this.attribute("size")
-        .flatMap((attribute) => {
-          const size = parseInt(attribute.value, 10);
-          if (size === size && size === (size | 0)) {
-            return Some.of(size);
-          }
-          return None;
-        })
-        .getOrElse(() => (this.attribute("multiple").isSome() ? 4 : 1));
-    }
-
-    return this._displaySize;
-  }
-
-  private _optionsList: Sequence<Element<"option">> | undefined;
-
-  /**
-   * {@link https://html.spec.whatwg.org/multipage/form-elements.html#concept-select-option-list}
-   */
-  public optionsList(this: Element<"select">): Sequence<Element<"option">> {
-    if (this._optionsList === undefined) {
-      this._optionsList = this.children()
-        .filter(Element.isElement)
-        .flatMap((child) => {
-          switch (child.name) {
-            case "option":
-              return Sequence.from([child as Element<"option">]);
-
-            case "optgroup":
-              return child
-                .children()
-                .filter(Element.isElement)
-                .filter(
-                  (grandchild): grandchild is Element<"option"> =>
-                    grandchild.name === "option",
-                );
-
-            default:
-              return Sequence.empty();
-          }
-        });
-    }
-
-    return this._optionsList;
-  }
+  protected _inputType: helpers.InputType | undefined;
+  protected _displaySize: number | undefined;
+  protected _optionsList: Sequence<Element<"option">> | undefined;
 
   /*
-   * End of methods specific to some kind of elements.
+   * End of caches for methods specific to some kind of elements.
    */
 
   /**
